Cleanup

+++ core/src/jvmMain/kotlin/info/laht/threekt/lights/LightShadow.kt
@@ -44,7 +44,7 @@ class SpotLightShadow : LightShadow<PerspectiveCamera>(PerspectiveCamera(50, 1f,
     fun update(light: SpotLight) {
 
         val fov = RAD2DEG * 2 * light.angle
-        val aspect = this.mapSize.width.toFloat() / this.mapSize.height
+        val aspect = this.mapSize.width / this.mapSize.height
         val far = light.distance
 
         if (fov != camera.fov || aspect != camera.aspect || far != camera.far) {
